Handle empty BT response result  CharacteristicReadOperation
In the current implementation, and empty response array in CharacteristicReadOperation.execute will be interpreted as a successful transmission (at this low level).

This is an exploration of how the code would be impacted if we were to introduce a `empty` state for this case, which would not be treated as successful at the level of CharacteristicReadOperation.